requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
alabaster==0.7.11         # via sphinx
apipkg==1.5               # via execnet
appdirs==1.4.3            # via black
argh==0.26.2              # via sphinx-autobuild, watchdog
atomicwrites==1.1.5       # via pytest
attrs==18.1.0             # via black, pytest
babel==2.6.0              # via sphinx
black==18.6b4
bumpversion==0.5.3
certifi==2018.4.16        # via requests
chardet==3.0.4            # via doc8, requests
click==6.7                # via black
coverage==4.5.1
doc8==0.8.0
docutils==0.14            # via doc8, restructuredtext-lint, sphinx
execnet==1.5.0            # via pytest-cache
flake8==3.5.0
idna==2.7                 # via requests
imagesize==1.0.0          # via sphinx
jinja2==2.10              # via sphinx
livereload==2.5.2         # via sphinx-autobuild
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==4.2.0     # via pytest
packaging==17.1           # via sphinx, tox
pathtools==0.1.2          # via sphinx-autobuild, watchdog
pbr==4.2.0                # via stevedore
pep8==1.7.1               # via pytest-pep8
pkginfo==1.4.2            # via twine
pluggy==0.6.0             # via pytest, tox
port-for==0.3.1           # via sphinx-autobuild
py==1.5.4                 # via pytest, tox
pycodestyle==2.3.1
pyflakes==1.6.0           # via flake8, pytest-flakes
pygments==2.2.0
pyparsing==2.2.0          # via packaging
pytest-cache==1.0         # via pytest-flakes, pytest-pep8
pytest-cov==2.5.1
pytest-flakes==3.0.2
pytest-pep8==1.0.6
pytest==3.6.3
pytz==2018.5              # via babel
pyyaml==3.13              # via sphinx-autobuild, watchdog
requests-toolbelt==0.8.0  # via twine
requests==2.19.1          # via requests-toolbelt, sphinx, twine
restructuredtext-lint==1.1.3  # via doc8
six==1.11.0               # via doc8, livereload, more-itertools, packaging, pytest, sphinx, stevedore, tox
snowballstemmer==1.2.1    # via sphinx
sphinx-autobuild==0.7.1
sphinx==1.7.6
sphinxcontrib-websupport==1.1.0  # via sphinx
stevedore==1.29.0         # via doc8
toml==0.9.4               # via black
tornado==5.1              # via livereload, sphinx-autobuild
tox==3.1.2
tqdm==4.24.0              # via twine
twine==1.11.0
urllib3==1.23             # via requests
virtualenv==16.0.0        # via tox
watchdog==0.8.3           # via sphinx-autobuild
wheel==0.31.1
